+++ eras/dijkstra/impl/cddl/lib/Cardano/Ledger/Dijkstra/HuddleSpec.hs
@@ -1014,6 +1014,11 @@ instance HuddleRule "protocol_param_update" DijkstraEra where
         , opt (idx 46 ==> VUInt `sized` (4 :: Word64)) //- "max endorser block txs size"
         , opt (idx 47 ==> huddleRule @"ex_units" p) //- "max endorser block ex units"
         , opt (idx 48 ==> VUInt `sized` (4 :: Word64)) //- "max ref script size per endorser block"
+        , opt (idx 49 ==> VUInt `sized` (8 :: Word64)) //- "peras minimum candidate block age in slots"
+        , opt (idx 50 ==> huddleRule @"positive_interval" p) //- "peras healing factor"
+        , opt (idx 51 ==> VUInt `sized` (2 :: Word64)) //- "peras certificate boost"
+        , opt (idx 52 ==> VUInt `sized` (2 :: Word64)) //- "peras target committee size"
+        , opt (idx 53 ==> (arr [] / arr [a (VUInt `sized` (8 :: Word64))])) //- "peras bootstrap round"
         ]
     where
       -- The maximum addressable 'LeiosSeatId' in bytes.