+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Examples.hs
@@ -63,6 +63,11 @@ import Cardano.Ledger.Dijkstra.PParams (
   ppMaxRefScriptSizePerBlockL,
   ppMaxRefScriptSizePerEndorserBlockL,
   ppMaxRefScriptSizePerTxL,
+  ppPerasBootstrapRoundL,
+  ppPerasCertBoostL,
+  ppPerasHealingFactorL,
+  ppPerasMinCandidateBlockAgeL,
+  ppPerasTargetCommitteeSizeL,
   ppRefScriptCostMultiplierL,
   ppRefScriptCostStrideL,
   ppuLeiosAnnouncementPeriodLengthL,
@@ -76,6 +81,11 @@ import Cardano.Ledger.Dijkstra.PParams (
   ppuMaxRefScriptSizePerBlockL,
   ppuMaxRefScriptSizePerEndorserBlockL,
   ppuMaxRefScriptSizePerTxL,
+  ppuPerasBootstrapRoundL,
+  ppuPerasCertBoostL,
+  ppuPerasHealingFactorL,
+  ppuPerasMinCandidateBlockAgeL,
+  ppuPerasTargetCommitteeSizeL,
   ppuRefScriptCostMultiplierL,
   ppuRefScriptCostStrideL,
  )
@@ -106,6 +116,7 @@ import Cardano.Ledger.State (
   BlsKey (..),
   StakePoolParams (..),
  )
+import Cardano.Slotting.Slot (SlotInterval (..))
 import qualified Data.ByteString as Strict
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
@@ -182,6 +193,11 @@ exampleDijkstraGenesis =
           , udppMaxEndorserBlockTxsSize = 12 * 1024 * 1024 -- 12 MiB
           , udppMaxEndorserBlockExUnits = OrdExUnits $ ExUnits 7_000_000_000 2_000_000_000_000
           , udppMaxRefScriptSizePerEndorserBlock = 12 * 1024 * 1024 -- 12 MiB
+          , udppPerasMinCandidateBlockAge = SlotInterval 90
+          , udppPerasHealingFactor = 1 %! 2
+          , udppPerasCertBoost = 15
+          , udppPerasTargetCommitteeSize = 800
+          , udppPerasBootstrapRound = SJust 0
           }
     }
 
@@ -353,6 +369,11 @@ exampleDijkstraOnwardsEraPParams =
     & ppMaxEndorserBlockTxsSizeL .~ 12 * 1024 * 1024
     & ppMaxEndorserBlockExUnitsL .~ OrdExUnits (ExUnits 7_000_000_000 2_000_000_000_000)
     & ppMaxRefScriptSizePerEndorserBlockL .~ 12 * 1024 * 1024
+    & ppPerasMinCandidateBlockAgeL .~ SlotInterval 90
+    & ppPerasHealingFactorL .~ 1 %! 2
+    & ppPerasCertBoostL .~ 15
+    & ppPerasTargetCommitteeSizeL .~ 800
+    & ppPerasBootstrapRoundL .~ SJust 0
 
 exampleDijkstraOnwardsEraPParamsUpdate ::
   (DijkstraEraPParams era, ConwayEraPParams era) => PParamsUpdate era
@@ -371,6 +392,11 @@ exampleDijkstraOnwardsEraPParamsUpdate =
     & ppuMaxEndorserBlockTxsSizeL .~ SJust (12 * 1024 * 1024)
     & ppuMaxEndorserBlockExUnitsL .~ SJust (OrdExUnits (ExUnits 7_000_000_000 2_000_000_000_000))
     & ppuMaxRefScriptSizePerEndorserBlockL .~ SJust (12 * 1024 * 1024)
+    & ppuPerasMinCandidateBlockAgeL .~ SJust (SlotInterval 90)
+    & ppuPerasHealingFactorL .~ SJust (1 %! 2)
+    & ppuPerasCertBoostL .~ SJust 15
+    & ppuPerasTargetCommitteeSizeL .~ SJust 800
+    & ppuPerasBootstrapRoundL .~ SJust (SJust 0)
 
 exampleBlsKey :: BlsKey
 exampleBlsKey =