+++ libs/cardano-ledger-binary/src/Cardano/Ledger/Binary/Encoding/EncCBOR.hs
@@ -59,6 +59,7 @@ import Cardano.Slotting.Slot (
   EpochInterval (..),
   EpochNo (..),
   EpochSize (..),
+  SlotInterval (..),
   SlotNo (..),
   WithOrigin (..),
  )
@@ -434,6 +435,8 @@ instance EncCBOR SlotNo where
 
 instance Serialise.Serialise t => EncCBOR (WithOrigin t) where
   encCBOR = fromPlainEncoding . Serialise.encode
+
+deriving instance EncCBOR SlotInterval
 
 deriving instance EncCBOR EpochNo